requirements.txt
amqp==2.2.1
billiard==3.5.0.3
bip32utils
bitcoin==1.1.42
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.33.0
django-appconf==1.0.2
django-celery-email==1.1.5
django-rest-auth==0.9.1
djangorestframework==3.6.4
eth_keys
ethereum==2.0.4
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
oauthlib==2.0.2
pbkdf2==1.3
pika==0.11.0
psycopg2==2.7.3
pycryptodome==3.4.7
pyethash==0.1.27
pysha3==1.0.2
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
requests-oauthlib==0.8.0
rlp==0.6.0
scrypt==0.8.0
six==1.10.0
urllib3==1.22
vine==1.1.4
pyotp==2.2.6
web3==5.4.0
django-extensions==2.2.6
python-binance==0.7.5
